trunks: added ScopedKeyHandle class
BUG=chromium:415999
TEST=manual and unit

Reviewed-on: https://chromium-review.googlesource.com/231391
Tested-by: Utkarsh Sanghi <usanghi@chromium.org>
Commit-Queue: Utkarsh Sanghi <usanghi@chromium.org>

+++ trunks/hmac_authorization_delegate.cc
@@ -22,7 +22,7 @@ const uint16_t kNonceMaxSize = 32;
 const uint8_t kDecryptSession = 1<<5;
 const uint8_t kEncryptSession = 1<<6;
 const uint8_t kLabelSize = 4;
-const uint16_t kAesIVSize = 16;
+const size_t kAesIVSize = 16;
 const uint32_t kTpmBufferSize = 4096;
 
 }  // namespace
