+++ trunks/hmac_authorization_delegate.h
@@ -17,8 +17,8 @@
 namespace trunks {
 
 const uint8_t kContinueSession = 1;
-const uint16_t kAesKeySize = 16;      // 128 bits is minimum AES key size.
-const uint32_t kHashDigestSize = 32;  // 256 bits is SHA256 digest size.
+const size_t kAesKeySize = 16;      // 128 bits is minimum AES key size.
+const size_t kHashDigestSize = 32;  // 256 bits is SHA256 digest size.
 
 /* HmacAuthorizationDelegate is an implementation of the AuthorizationDelegate
  * interface. It provides the necessary Auth data for HMAC sessions.
